Store save state as PNG
mGBA supports two ways of storing save states:
(a) as a raw binary file,
(b) as a PNG file with the (compressed) state data embedded in a custom chunk

So far, the bot has created the former because that's just easier to do.

This changes it so it creates PNG files. mGBA can load both, so only the saving logic needed to change.

Benefits for this different approach:
- File can be viewed in any image viewer and will show a thumbnail of the bot state,
- Because the save state is compressed, the file size for each state file went from ~390 kB to ~50 kB
- If I ever get around to implementing a 'Select state to load' screen, we could use that embedded image as a preview thumbnail.