Fix debug tabs in non-Emerald games
The Symbols tab was crashing on (at least) Ruby/Sapphire because one of the default symbols (`sChat`) does not exist in their symbols table.

I fixed that by automatically removing any symbol that could not be found, rather than throwing an Exception.

Also, the Daycare tab broke on Ruby because the case where `ParsePokemon()` returns `None` was not handled correctly.