Windows: Show 'Press Enter to continue' prompt even during very early errors on Windows
At the moment, the `atexit` handler that is supposed to keep the console window open on close is registered in the GUI source file.

But if there is an error before that file could be loaded (usually because `requirements.py` wasn't run properly) then the console window will disappear immediately and the user will not see the error message.

In order to make support a bit easier, this registers the `atexit` handler as soon as possible.