Check and install requirements in main `pokeboy.py` file
This removes `requirements.py` as a separate file and inlines it in the main `pokebot.py` file.

The idea being that it should be impossible to run the bot without the requirements installed.

To avoid `pip` being run on every bot start and thus increasing startup time, after successful loading of all requirements the current bot version is stored in `.last-requirements-check`.

On startup, the bot first checks whether the current version is equal to the one stored in that file, and if so it skips all the checks.

That means you can force reloading the requirements (for dev purposes, probably) by just removing that file.

In addition to the regular requirements, it also checks that
- a 64-bit version of Python is running
- a supported version of Python is running (currently 3.10 or 3.11)

Hopefully, this will avoid some of the common issues in the support channel.

If it does, I expect that everyone working as supporter chant my name before going to sleep.