Add a basic GitHub Actions workflow for releases
This adds a GitHub Actions workflow that will automatically create a release whenever a tag is pushed/created.

While doing so, it will update `modules/version.py` to contain the name of the tag as version number, and also create a little changelog in the release description.

If the original source code from the Git repository is run, the version number will either be generated by the current commit hash (e.g. `dev-a1b2c3d`) or just be 'dev' if no Git information could be found.

Apart from that, this changes a couple of things to improve pyinstaller packaging.

Unfortunately, using pyinstaller bundles for distribution is not really going to work because Windows Defender will just reject those .exe files because they are unsigned.

Signing them would require a certificate that costs a couple hundred USD per year, so that's not really worth it. And explaining to people that they should click to ignore the warning seems questionable at best, and would probably mean at least the same amount of questions in the support channel.

Also, I have extracted the requirements-related stuff into a `requirements.py` again, to clean up the code in `pokebot.py` a bit and to allow manual triggering of a requirements fetch.

`pokebot.py` (and now also `import.py`) will still call the functions therein, so people don't _have to_ run `requirements.py`.