Extract Pokemon Storage System data
This adds functionality to access the Pokemon Storage System (the PC thing) and extract all Pokemon stored in there.

It also adds a check to the importing feature whether the Pokemon already exists in the system. This prevents a Pokemon from being imported multiple times if the bot is closed and restarted in the middle of a battle.