Add an auto updater
This change adds a very crude self-update system to the bot that checks whether the most recent GitHub release is newer than the version currently running.

It only does so once a day, and only if there is no `.git` directory present (which would indicate that the user uses a VCS to manage the version.)

The updater can be manually triggered by running `updater.py`. But when using Git this might _downgrade_ the current version, as it would just install the most recent release.

If the user allows the update (there is a confirmation prompt, of course) it will just download the release ZIP and extract it into the current directory. It will not remove any old files, though that shouldn't usually be an issue.