Add basic success check to `write_file()` and use atomic file-replacing function
This checks whether `write()` returns the correct number of characters written, and uses `os.replace()` to replace the file (which is supposed to be atomic) rather than deleting the old file and _then_ renaming the new one.

This may or may not help with instances of file corruption what we have seen. Since it's difficult to reproduce the issue, it's hard to be sure about that.

But at the very least it shouldn't make things worse.