Refactor Modes
This is an attempt to change the interface for modes a bit, so they make more use of Python's Generator features and allow easier reuse of common functionality, such as soft resetting or navigating around.

I have focussed particularly on the Starters mode, which works essentially the same as before (with some tweaks) but looks a bit different in terms of its code.

As a major tweak to the Starters mode I have added functions to read the current save game, so we can verify that (a) there is one to being with, and (b) it is saved in one of the locations required by the starter modes. Before, it would only check whether the player was currently standing in front of the starter bag, without being sure that they saved in that place.