Update flag names and add search field to debug tab
This updates the list of flags with ones from the decompilation project, which seem to be a bit more complete in some places.

Also, the most recent versions of the MissingEventFlagsCheckerPlugin renamed a lot of the flags and apparently migrated towards human-readable names, so updating those wouldn't have been that easy anyway.

Also, @pekempy requested a search field in the Event Flags tab which I've also implemented. There are some quirks with that (ordering is messed up when searching, and keystrokes still get registered by the emulator) but it should be _good enough_.