Add an HTTP endpoint for rendering the entire world map
This adds a `/world_map` endpoint that tries to render a mini map of the map the player is currently standing on, as well as all connected maps.

So if the player stands somewhere outside, this should more or less render the entire region as all those maps are connected to each other.

It also displays all 'object templates' (which are entities in the world that _might_ exist, depending on a flag -- the flags aren't checked, so this view will show _all_ potential objects.)

The page isn't updated as the player moves around and generally isn't all that useful, but it serves as a bit of a test case for various map-related functions.

And it looks pretty.