Fix handling of dynamic warps in map code
This adds handling for some special cases related to map warps.

Dynamic warps are warps that have a dynamic destination based on some in-game state.

One example for that is the entrance area of Terra Cave on Emerald (where Groudon is found.)

Before, the Map debug tab would crash the emulator on this map. Now it works.