Add lookup functions for map encounters
This adds a `get_wild_encounters_for_map()` function that, as the name implies, will return the list of encounters (grouped by categories such as land, surfing, fishing) for a given map.

That information is displayed in the `Map` debug tab and not used anywhere else so far, but might prove useful eventually.

This commit also removes some 'cheatey' functions that I have used for testing in the past and forgot to remove after they were no longer needed.