Fix GIF playback speed and some other stuff
This sets the GIF 'frame rate' to a static 20ms/frame instead of using the number of frames as the number of milliseconds per frame.

The GIF generation is a bit faster now by not using `peek_frame()` for every single frame-to-render but rather using one peek pass for all of them.

It also changes `peek_frame()` to create a context so it can be used by invoking the `with` keyword which might make the code a bit more readable.

This also fixes a bug in the Sudowoodo mode that caused an error message whenever you open the bot mode menu on FR/LG. That doesn't have anything to do with GIFs, but I thought I'd sneak it in anyway.