Allow disabling GUI theme with an environment variable
The new theme makes the GUI laggy to the point of being unusable.

This change allows disabling the theme using an environment variable (starting the bot with `POKEBOT_UNTHEMED=1`) so I can more happily work and test on the bot.

This changes some of the type hints back from `ThemedTk` to `Tk`, as the former is a subclass of the latter so it works either way.