Error message due to Nugget Bridge mode in RSE
The Nugget Bridge mode's `is_selectable()` method does not define the `allowed_maps` on RSE, but it _does_ try to compare the current map to it.

This leads to an error being thrown each time the bot mode menu is opened.

The error is suppressed in non-debug runs so that is not a huge issue for most users, but still.