Do not replace Unown species name
Currently, the bot _replaces_ the name of the global species object for Unown by one that contains the letter of the last-encountered Unown.

This is done in the constructor of every `Pokemon` object, and I assume this change was supposed to make the Unown letter show up in the stats.

This approach comes with some drawbacks, though:

(a) Since the `Species` objects are global, calls to `get_species_by_name("Unown")` might unexpectedly fail if an Unown has been encountered before (or is in the player's party.)

(b) It will crash the bot whenever it encounters an invalid Pokémon (such as when a battle starts and the party is briefly invalid for a single frame) because it tries to access the species of an invalid Pokémon.

(c) Since this is done in the constructor, it will force Pokémon data to be sorted and decrypted even if that Pokémon's data is never actually accessed (for example when you call `get_party()` but only look at the first entry while there are other Unowns further back.) This might introduce a (likely very mild) performance impact.

So this change just adds a dedicated property for the species/Pokémon name that is used by the stats system instead.