Create a unified check whether player avatar is controllable
Currently, there are a lot of places where we are checking for the `heldMovementActive` or `heldMovementFinished` flags, or some other indicators that the player avatar is currently controllable.

I've now migrated that into a single function that does all the checks in order to make the code a bit easier to understand and also to avoid inconsistencies across modes.

While doing so, I've extracted `get_player_map_object()` into its own function, because that's what _most_ of the calls to `get_map_objects()` actually want to know.

This fixes #249 and fixes #259