Prevent overshooting the target when walking
The `walk_to()` utility function keeps the direction buttons pressed until the player avatar reached its destination.

This leads to issues for example with the Repel listener (but possibly with others too, like the Pokenav listener) where it might end up running one tile too far because the buttons are not being released in time.

This changes the `walk_to()` function so that the buttons are only pressed when the avatar is not moving, or has just reached a tile (`CENTERING` state.)

Fixes #253