Fix Spin mode on underwater maps
The battle listener will, at the end of a battle, spam the `B` button until the player is controllable again.

Pressing a button continuously means that it will be pressed in one frame, then not pressed in the next, then again etc.

So there is a 50/50 chance that during the frame where the player avatar becomes controllable again, the button is being held down.

This appears to happen regularly after a battle, so on Underwater maps this triggers the surfacing dialogue.

I have added an exception for this very case.