Do not go into an infinite loop when trying to rotate the lead in a trainer battle
If the bot is configured to rotate the lead Pokémon when it gets below a certain HP threshold (the `hp_threshold` setting in `battle.yml` with `lead_cannot_battle_action` set to `rotate`), the game goes into an infinite loop and appears to crash.

This only happens in trainer battles because in wild encounters it would just try to flee.

But in trainer battles, it will set the action to `switch` and the Pokémon to switch to as `None` -- which in some part of the code is supposed to trigger an error that puts the bot back into manual mode, but it never reaches the code because before that, this loop is running:

```python
while self.choice is None or self.idx is None:
    self.determine_battle_menu_action()
```

Since `self.idx` (the Pokémon to switch to) will always be set for `None` in the aforementioned case, the bot can never break out of this loop.