Nugget Bridge: Check for `hp_threshold` and use new navigation utility
This updates and simplifies the Nugget Bridge mode by making use of the new `navigate_to()` function.

While implementing this, I've also added some checks to that function so that it better recognizes when a script is triggered and does not get stuck by that.

Also, since the mode effectively requires the `hp_threshold` feature to be disabled (set to `0`) because otherwise it might just get stuck when your sacrificial Pokémon gets too weak, I have added a check for that.