Stepping Mode: Allow stepping _backwards_, add key controls
This change adds a buffer to the stepping mode that will keep track of the last 128 frames that have been stepped through, and allow the user to revert back to them.

That can be very useful in cases where you stepped one frame too many by accident, or just want to try how a certain event behaves with different key inputs.

Previously, you would have to recreate the situation each time or awkwardly work with save states.

In addition to that, this change adds keyboard controls to stepping mode:
- **Space** will advance by one step
- **Ctrl+Space** will revert one step

Fixes #70