Make fetching the player's location more reliable
When the player transitions onto another map, it seems like the avatar data (which includes the current map) can be corrupt.

We've already seen a similar issue when fetching the player's party, where especially during the start of a battle the data can be corrupt for one frame.

This change fixes the issue in the same way, by catching the error once and using the `peek_frame()` function to get the _next frame's_ map data.

Fixes #289 (I hope)