Daycare mode: Use full Mach Bike speed, fix R/S issues (#294)
This change updates the Daycare mode to make use of the new navigation functions.

These functions hit the direction buttons at different times, so that the player will actually go at full speed using the Mach Bike instead of looking like a grandpa.

I've also adjusted the path the player takes while waiting for a new egg to be available/to hatch to be a bit more efficient: When turning 180° on the Mach Bike, it will first come to a stop (which takes 3 tiles) and then slowly pick up speed again in the other direction. On the other hand, when turning 90° it does not lose any speed.

So going in a circle rather than back and forth is actually faster. It also means that we can stay closer the daycare man/house, meaning that we likely also reach them more quickly once it's time to pick up a new egg or release Pokémon.

Finally, while working on this I noticed that the code that pre-loads map data (introduced in #287) completely breaks navigation on R/S because some maps from the `MapRSE` enum do not actually exist in these games. I have added a check for that.