Roamer Reset mode: Check game state on FR/LG (#295)
This adds some checks on FR/LG that the in-game conditions for doing the Roamer run are actually met, i.e. that the player has the Sapphire in their inventory and that the process has not yet been started.

(While it would be enough to just check for the existence of the Sapphire item in the player's inventory, I'm checking for several vars/flags to give a more specific error message.)

While doing that, I have added some more helper methods to the `SaveData` class so we can check for flags in the saved game rather than the currently loaded one. (Since we're soft resetting immediately, we don't really care about the current flags.)