Static Gift Resets: Properly handle egg hatching
At some point, probably when the listeners were introduced, the SGR's mode handling of egg hatching broke.

Because the listener will intercept the egg hatching cutscene, the mode does not actually learn about it.

So the player would just keep riding their bike in a circle even though the gift egg had already hatched.

This is a fix for that.