Fix Puzzle Solver modes using bicycle
Sky Pillar and Mirage Tower rely on the Mach Bike going fast enough to outrun the cracked floor tiles breaking.

That does not work with the old navigation function because it hits the buttons at the wrong time, so cycling is very slow.

This change updates the Puzzle Solver mode to use the new navigation functions.

It also fixes an issue with replenishing repel, which probably affected Roamer Reset mode.