Update Roamer mode to use new navigation functions
This removes calls to the deprecated same-map navigation function in favour of the new cross-map one.

Doing that simplifies the code a little bit, and hopefully also fixes #316 (at least I wasn't able to reproduce the behaviour anymore.)

As a semi-related fix, this resolves an issue where reapplying Repel would lead to the player continue running in the same direction until they hit an obstacle.