Fix Safari Zone Rock Smash getting stuck when walking
When using the Rock Smash mode in Emerald's Safari Zone (for hunting Shuckle) and not using the bicycle (peasant!) the bot would get stuck after a few rounds.

This is probably due to in in-game bug (pointed out by @Terasol!) where the scripting context did not always get cleared after using rock smash.

A script that was originally meant to handle the Safari Zone entrance sequence would then wait for the script to 'finish' -- which would never happen, thus stalling the bot.

Fixes #314