Kecleon: Add to SSR mode, update Wiki
The Kecleon mode can only be used on Emerald (because the method itself doesn't work on R/S) but the Wiki still listed R/S as compatible with that mode.

As an alternative for R/S, I have added Kecleon to the Static Soft Resets mode. It's probably slower than the Whiteout method of the Kecleon mode, but it should work on R/S.