Suppress 'unsupported platform' error message for Desktop Notifications on Windows 11
The `notifypy` does not support Windows 11 and will raise an Exception when used on that platform.

Showing that error does not really help anyone -- we know about that issue, there's nothing we can do about it, and it just means that desktop notifications (a mostly cosmetic feature) won't work.

So rather than spamming confusing error message, this will just silently ignore the error.