Fix player running one additional tile before re-applying Repel
The repel listener had an issue where it would close the notification message about repel expiring, but then yield control back to the bot mode for one extra frame before attempting to re-apply another repel.

If that happened in or right next to tall grass, that would mean that the player moved inside the grass and might lead to an unexpected encounter.

The bot would then hang because it would keep spamming `Start` to open the menu.