Do not avoid weather-changing tiles when pathfinding
The pathfinding algorithm attempts to avoid tiles where a script could trigger.

This leads to an issue with the Kecleon mode where the player needs to cross some tiles that will change the weather. There is currently no handling for this special event type, so the bot tries to avoid those tiles and will inadvertently run into tall grass in the process.

I have added a detection for this event type, improved the debug map view and exempted those scripts in the pathfinding function.

Fixes #331