SGR: Use different path for Togepi egg hatching to avoid wandering NPC
When using the Static Gift Resets mode in FR/LG to reset the Togepi egg, there is an NPC on the island that may wander around.

The path programmed for cycling in a circle to hatch the egg intersected that NPC's path so the player would get stuck pretty quickly.

This just chooses a different area for cycling in.