Add lookup function whether a trainer has been defeated
This adds a lookup method to `ObjectEventTemplate` to see whether a trainer (map object) has already been defeated.

Since this is only available for `ObjectEventTemplate` and not for `ObjectEvent` itself, I have added a lookup function to get the former from the latter.

This has no immediate use (though it could be used in some modes to verify that relevant trainers have been defeated) but **chesta** kindly provided that code in the `development-chat` Discord channel.

I have added a 'defeated' indicator to the object list in the `Map` debug tab.