Add Berry Blender mode and support looking up Pokéblock data
This is a little offshoot of my battle system overhaul, as I needed a full Pokéblock case in order to test Safari zone battle functionality. PKHeX does not support editing Pokéblocks, so it had to be done in-game.

Obviously, writing this mode took _way_ more time than just manually filling up the case would have, but oh well.

It does nothing too fancy, just asks the user which berry to use and then runs the game once with perfect accuracy.

This also adds low-level lookup functions for Pokéblocks and adds the information to the `Player` debug tab.