Fix navigation getting tripped up by turning around at the start
When starting to walk while originally facing a different direction, the player avatar actually does _two_ steps: First turning around, then moving to the next tile.

If the tile the player is standing on can have encounters, this means for a single navigation step there are actually two opportunities for an encounter to happen.

This trips up the navigation system because it waits for the player to reach the next tile and will eventually re-attempt the navigation. Since a battle would have started at that point, navigation is not possible and an error will be thrown.

This was triggered in Roamer Reset mode **if the leading Pokémon did not have the Illuminate ability**. That's because _with_ the ability we do 18 turns (which means we face South at the end) while _without_ it we do 36 turns (facing North at the end.)

I have also adjusted that turn number.

Fixes #344