Fix broken startup if dependencies are not installed
When running `pokebot.py` for the first time, new users would likely get a `ModuleNotFoundError` because some dependencies are missing (such as `PIL`.)

That's because the new `PokebotHeadless` module/class is imported too early -- at the top of `pokebot.py` -- which means that Python tries to resolve the dependencies before the `check_requirements()` call even happens.

This moves the import to a later, safe point.