Make sure that all inputs are tracked by HTTP event stream
Adds a buffer for input events so that each time the HTTP server sends an `Inputs` event via the event stream (every 1/30th of a second) it is guaranteed to contain _all_ buttons that have been pressed since the last update.

That means that the event does not necessarily report the exact state of the buttons but more of an 'echo'.

For example, when using the Spin mode unthrottled, the event will likely report all 4 directional buttons to have been pressed at the same time.

But I think that's a good tradeoff between accuracy and performance.

I've also changed the previously introduced mechanism to wait for the `work_queue` by using an `Event` to using the built-in `task_done()` and `join()` methods of the queue object. Just found those by accident in the Python docs.