Fix encounter logging issues with breeding and soft resetting
This fixes two issues:

Affecting **Static Gift Resets with Togepi and Wynaut:** When hatching an egg, the `EggHatchListener` is logging the encounter. But the SGR mode itself also logs the encounter after 'looking' at the summary screen before resetting. So the encounter got logged twice.

I've fixed this by (a) adding a missing check in `log_encounter()` whether the Personality Value of an encounter matches the previous one, and (b) removing the encounter-logging from `EggHatchListener` and have the mode do that. I've updated Daycare mode as well, the only other one that does hatching.

And then, affecting **Static Soft Resets** and **Sudowoodo** modes (and possibly others): When resetting the emulator, the `BattleListener` would not reset itself and continue thinking that it's mid-battle.

Thus, it would never register another encounter until a battle actually _ended_, which in these modes never occurs.

I've fixed this by letting the listener compare the emulator's frame counter. If it's suddenly _less_ than before, it resets its state.

(Also fixes a syntax warning in the HTTP server.)