Fix handling of fainted Pokémon in non-English games
In non-English versions of Emerald, there was a symbol missing and because of that the bot didn't detect that it was already _past_ the selection of a new lead.

So it tried to start the faint handling again and got stuck in a loop.

Also, there was an issue with `TurnAction.use_item_on(item, party_index)`, where it wouldn't map the party index to the in-battle party order. That meant that it sometimes chose the wrong Pokémon to use the item on.