Fix TCG card, Starter mode, shiny reporting on console
Collection of smaller fixes:

(1) **Starters** mode on R/S/E no longer checks for both Pokémon cries to happen. This wasn't the case if the opponent Poochyena/Zigzagoon happened to be shiny and in that case the bot would get stuck. No more.

(2) A shiny encounter is now reported to the console _before_ it is logged by the stats system. That means that all the old shiny phase data is still available at that point, rather than being completely empty. It also means that the shiny's stats aren't included in the totals, but I guess that's OK.

(3) Added a failsafe in the `get_clock_time()` to report a zeroed-out time on FR/LG where there is no RTC system, rather than unceremoniously crashing.

(4) TCG card generation was broken because we are now first saving the image in a temporary file with a `.tmp` extension, which meant that PIL didn't know what format to use and silently failed.

I've added error reporting to the TCG card generation and fixed that particular issue.

(5) Possibly related to that, the Discord plugin had a mistake where it would indefinitely wait for an image file to appear. I intended there to be a timeout, but that was never actually checked.

This meant that if TCG card generation was enabled, it would just get stuck after the first shiny because it would wait forever for that card image to become available.