Display offset data for script context stacks (#424)
This adds information about pointers and offset from symbol to the script stack in the debug tabs to help in finding symbol locations for non-English games.

I've also changed the formatting for task data so the data payload is now displayed as 2-byte groups. That makes more sense since when using `task.data_value(n)` you get a 16-bit value.