Don't rotate lead in Level Grind mode if only lead is selected to be levelled
If the user chooses the 'level lead only' option, we use the default battle strategy -- but depending on configuration, this might rotate the lead after a battle if the previous lead was too weak.

This would mess up that mode because now a different Pokémon is being levelled.

I've added a modified battle strategy for that.