Make the bot at least run on Python 3.13
Updates some dependencies so we get versions that are compatible with Python 3.13.

This wasn't possible for all versions: For `aiortc` (used for providing a WebRTC-based video/audio stream via the HTTP server) there is no compatible version yet.

Since that is a rather niche use case, I have modified `requirements.py` to just not install that library if the user's Python version is too new. The video stream won't be available in that case.

I also modified the script slightly so that if the user's Python version is _newer_ than the newest version we have marked as compatible, it will offer a 'do it anyway' choice.