Fix re-disabling video after generating a GIF
When trying to generate a GIF with video disabled, the bot will enable video for the duration of the image generation and then disable it again.

That works fine for the enabling part, but the disabling call happens inside another thread which can lead to issues with Tkinter (`RuntimeError: main thread is not in main loop`.)

This change makes it so that the video disabling is done by the main thread.