Add `--user` to `pip` flags when not running in a virtualenv
By default, `pip` tries to install packages globally -- which leads to permission issues if Python is not installed in a user directory and the requirements script not run as superuser.

This changes the requirements script to it adds the `--user` flag to `pip` so it tries to install the packages only for the current user.

Fixes #420