Simplify logging config
This removes most configuration options for logging so that it can essentially only be turned on or off. This simplifies both the code and testing.

Since console logging can now only be turned on or off, it will always display _all_ available information. To save space, I've updated the output format a bit:

- Moves are colour-coded by type, but there is no more additional information such as type name, PP, accuracy etc. People can look that up on Bulbapedia if they're interested, no value in displaying that here.
- EV Yield is displayed (could also be looked up, but might at least be vaguely useful here.)
- The IVs table contains and up/down arrow behind the stat that is increased/decreased by the Pokémon's nature.