Take abilities into account when calculating effective encounter rates
When calculating effective encounter rates on a map, this will check whether the abilities Static, Magnet Pull, Intimidate or Keen Eye are active, all of which may modify the proportion of encounters.

I've also added an info bubble about encounter-relevant abilities to the stream overlay and fixed some timezone-related issues.