Fix 'Total Encounters' value in Discord messages
Changes the Discord integration plugin so it uses the `GlobalStats` object rather than the current phase's snapshot values (which are only available when there's a Shiny encounter.)

I also moved where the plugins are loaded to the main loop, because at that point the selected `Profile` is available.

This is relevant because the built-in plugins (Discord, OBS, GIF/TCG Generator) only load if configuration actually uses it. So if these things were disabled globally and only enabled in the profile-specific config, they wouldn't load.