Add Wiki pages for plugins and the stats database
This adds some basic information about how plugins are supposed to work, and also extends the base `BotPlugin` class with comments for each function.

Until now, a plugin file was only allowed to contain _one_ plugin class -- but there's no real reason why it has to be this way, so I've removed that restriction.

I've also added a very brief description of the contents of the stats database to the wiki.