Fix plugin loading again
I found out the reason why `load_plugins()` was called very early on in `pokebot.py`: Because that's the point where the bot mode list is queried.

So right now, a mode added through a plugin would never show up in the list because it's being loaded too late.

I've split up loading user-provided plugins (early) and built-in plugins (only after profile selection so that profile config is available.)