Rock Smash: Check correct party member for Repel level, describe better strategy in Wiki
Repel works by using the level of the first **non-fainted, non-egg** Pokémon in the player's party.

But both Rock Smash and Roamer Reset modes just checked the first party slot. So if the first Pokémon in the party was fainted, it would erroneously still use that for the repel level check.

Also, I have updated Rock Smash's Wiki page to advise users to use a Pokémon with the Vital Spirit or Pressure abilities as their first party member, as that gives another +35% boost to encounter rates.

This possibility was found by **14flash** on the Discord server and verified through benchmarking.