Take previously fainted party members into account when rotating lead
The default battle handler would, until now, always check the first Pokémon in the party when deciding whether to rotate the lead after a battle -- even if that first Pokémon was unable to battle (fainted, egg) before the battle already.

This means that when having a fainted Pokémon in that first slot on purpose (in order to make use of its ability), the bot would still rotate the lead and thus disable the ability.

This change makes it so the bot remembers the first non-fainted lead before a battle, and ignores any party slots before that.