Defer loading of keyboard data until it is needed
The `keyboard` module will attempt to load the character table for the in-game keyboard as soon as the battle handler is being imported anywhere (since the battle handler will import the module, thus immediately executing the loading code.)

So if the battle handler is imported anywhere before a profile has been loaded (such as importing it in a plugin), the bot will crash because `context.rom` is still `None`.

This defers loading the language-specific profile until it is actually needed.

I've also converted the dictionary structure into a dataclass to get some autocompletion while coding.