Prevent unnecessary status moves when catching
The Catch battle strategy currently tries to inflict Sleep or Paralysis on an opponent to increase the odds of catching it.

But in cases where the likelihood of a Pokémon being caught in one attempt is >= 50% we might as well just throw balls immediately as that would use the same amount of turns or less.

So this PR...

- adds a helper function to calculate the odds of a successful catch,
- uses that function to only use sleep/paralysis-inducing moves if the chance is <50%,
- makes sure we do not attempt to use status-inflicting moves when our Pokémon is under the influence of Taunt (which would prohibit it from using non-damaging moves.)