Add plugin hook for config reload
This allows plugins to be notified when the bot config is reloaded (`Ctrl+C`.)

Might be useful if plugins calculate some stuff based on configuration but do not always want to check whether that config has changed.

But I need it specifically to allow the Stream Plugin to have its own configuration and also reload that on demand.