Allow specifying the spin direction
This adds a parameter to the `spin()` utility function that controls whether we're spinning clockwise (default and the way it previously worked) or counter-clockwise.

This is not used by the bot as is, but can be used by plugins -- planning to use this as a very low-key easter egg in the Stream plugin.