Fix Rock Smash for R/S
This fixes the Rock Smash mode getting stuck waiting for a script that just has a different name in the R/S decompilation.

I've confirmed it to work in Granite Cave, but since I don't have a test profile for the Safari Zone I did not update the wiki yet to reflect that support.