Fix Rock Smash still getting stuck on R/S
When getting an encounter through Rock Smash on R/S, the script context never turns off -- which means that the mode gets stuck waiting for exactly that to happen.

This changes the exit condition to the rock's flag changing, which worked reliably for me throughout an in-game hour of running the test.