Stats DB: Improve error message for Database Locked and Unique Constraint Violation (#527)
This adds more verbose error messages for two common issues related to running multiple instances of a profile:

`Database is locked` error, which can happen at pretty much any point if two or more instances are open.

`Unique Constraint Violation`, which can happen when logging an encounter if the encounter ID has already been used.

The bot reads the highest encounter ID on startup and then just increments that number internally. But that means that if you ran two or more instances of the profile, each instance would get the same initial encounter ID and then try to use that +1. The second instance to do so will get the Unique Constraint Violation error because the DB does not accept duplicate values for the Primary Key.