Add support for Surf and Waterfall in pathfinding
This updates the `calculate_path()` function as well as `navigate_to()` to be able to start/stop surfing, as well as climb and swim down waterfalls.

As a proof of concept, I have updated the Level Grind mode to support grind spots on water. This is probably not all that useful, but it served as a nice test case.