Fix cost calculation in pathfinding
The latest improvements to the pathfinding algorithm (Surf, Waterfall, and Acro Bike support) have introduced a bug where the cost of a route wasn't properly calculated.

Instead of accumulating the cost of the entire route, for each tile the cost would be reset to 1 again.

The pathfinding would still _work_, but it would no longer consistently find the most optimal route and instead go on extended sightseeing trips through towns and routes.