Fix error when trying to create a new Pokémon in the party edit menu on Japanese
This fixes an issue where the `encode_string()` function (used for encoding the nickname when editing the party through that new debug tool) would always use the international character table.

This caused issues when creating a Pokémon without a nickname in a Japanese game, because then by default the nickname would be set to that language's species name.

I've also extracted the Pokémon-creating and party-writing functions into `debug_utilities.py` to make the `debug_edit_party.py` a bit shorter, and to only have GUI code in there.