Add debug tools for adding a test party and test item bag
This adds two more items to the `Debug` menu, one for overwriting the current item bag with a 'testing toolkit' that contains most items that could become useful at some point.

The other one adds three Pokémon to the party: A strong Mewtwo for battling everything, a Lotad that knows moves useful for catching, and two 'HM slaves' that know all the HM moves.