Refactor party code
This creates a dedicated container class for a party of Pokémon, which is used for the return value of both `get_party()` and `get_opponent_party()`.

It makes it a bit easier to handle by adding some convenience lookup functions to that class, rather than having to do the same 'iterate over party to see if some condition is met' pattern all over the code base.

There's probably room for more improvement, but for now I've added lookup functions for non-fainted Pokémon, eggs, and Pokémon that know a particular move.